refactor recipe UI into components
