+++ src/Recipes.jsx
@@ -3,33 +3,24 @@ import React, { useEffect, useState } from 'react'
 import {
   collection,
   onSnapshot,
-  addDoc,
-  updateDoc,
-  deleteDoc,
-  doc
+  addDoc
 } from 'firebase/firestore'
-import { db, storage } from './firebase'
-import { ref as storageRef, uploadBytes, getDownloadURL, deleteObject } from 'firebase/storage'
-import './Recipes.css'
+import { db } from './firebase'
 import { query, orderBy } from 'firebase/firestore'
+import './Recipes.css'
+import RecipeCard from './components/RecipeCard'
+import RecipeOverlay from './components/RecipeOverlay'
 
 export default function Recipes() {
   const FAMILY_ID = 'sharedFamily'
   const colRef = collection(db, 'families', FAMILY_ID, 'recipes')
   const q      = query(colRef, orderBy('title'))     // tri alphabétique par titre
 
-  const [recipes, setRecipes]         = useState([])
-  const [searchTerm, setSearchTerm]   = useState('')
-  const [selected, setSelected]       = useState(null)
-  const [editMode, setEditMode]       = useState(false)
-  const [draftTitle, setDraftTitle]   = useState('')
-  const [draftIngredients, setDraftIngredients] = useState([])
-  const [draftSteps, setDraftSteps]   = useState([])
-  const [draftNotes, setDraftNotes]   = useState('')
-  const [draftImageUrl, setDraftImageUrl] = useState('')
-  const [imageFile, setImageFile] = useState(null)
-  const [pasteMode, setPasteMode]     = useState(false)
-  const [pasteText, setPasteText]     = useState('')
+  const [recipes, setRecipes]       = useState([])
+  const [searchTerm, setSearchTerm] = useState('')
+  const [selected, setSelected]     = useState(null)
+  const [pasteMode, setPasteMode]   = useState(false)
+  const [pasteText, setPasteText]   = useState('')
 
   // real‑time load
   useEffect(() => {
@@ -52,21 +43,9 @@ export default function Recipes() {
     r.title.toLowerCase().includes(searchTerm.toLowerCase())
   )
 
-  // open overlay + init drafts
+  // open overlay
   function openOverlay(r) {
     setSelected(r)
-    setEditMode(false)
-    setDraftTitle(r.title || '')
-    setDraftIngredients(
-      (r.ingredients || []).map(i => ({
-        name: i.name || '',
-        quantity: i.quantity || ''
-      }))
-    )
-    setDraftSteps(r.steps || [])
-    setDraftNotes(r.notes || '')
-    setDraftImageUrl(r.imageUrl || '')
-    setImageFile(null)
   }
 
   // add new recipe
@@ -80,89 +59,6 @@ export default function Recipes() {
       notes: '',
       imageUrl: ''
     })
-  }
-
-  // save all edits
-  async function saveAll() {
-    if (!selected) return
-    const ref = doc(db, 'families', FAMILY_ID, 'recipes', selected.id)
-    let url = draftImageUrl
-    try {
-      if (imageFile) {
-        const imgRef = storageRef(storage, `recipes/${selected.id}`)
-        await uploadBytes(imgRef, imageFile)
-        url = await getDownloadURL(imgRef)
-        setImageFile(null)
-        setDraftImageUrl(url)
-      } else if (!draftImageUrl && selected.imageUrl) {
-        const imgRef = storageRef(storage, `recipes/${selected.id}`)
-        await deleteObject(imgRef).catch(() => {})
-      }
-    } catch (err) {
-      console.error('Image upload failed:', err)
-    }
-    await updateDoc(ref, {
-      title: draftTitle,
-      ingredients: draftIngredients,
-      steps: draftSteps,
-      notes: draftNotes,
-      imageUrl: url
-    })
-    // reflect locally
-    setSelected({
-      ...selected,
-      title: draftTitle,
-      ingredients: draftIngredients,
-      steps: draftSteps,
-      notes: draftNotes,
-      imageUrl: url
-    })
-    setEditMode(false)
-  }
-
-  // delete current
-  async function deleteCurrent() {
-    if (!selected) return
-    if (confirm('Supprimer cette recette ?')) {
-      await deleteDoc(doc(db, 'families', FAMILY_ID, 'recipes', selected.id))
-      if (selected.imageUrl) {
-        await deleteObject(storageRef(storage, `recipes/${selected.id}`)).catch(() => {})
-      }
-      setSelected(null)
-    }
-  }
-
-  function removeImage() {
-    setDraftImageUrl('')
-    setImageFile(null)
-  }
-
-  // ingredient helpers
-  const setIngName = (i, v) => {
-    const arr = [...draftIngredients]; arr[i].name = v; setDraftIngredients(arr)
-  }
-  const setIngQty = (i, v) => {
-    const arr = [...draftIngredients]; arr[i].quantity = v; setDraftIngredients(arr)
-  }
-  const addIngredient = () =>
-    setDraftIngredients(arr => [...arr, { name:'', quantity:'' }])
-  const removeIngredient = i =>
-    setDraftIngredients(arr => arr.filter((_, idx) => idx !== i))
-
-  // step helpers
-  const setStep    = (i, v) => {
-    const arr = [...draftSteps]; arr[i] = v; setDraftSteps(arr)
-  }
-  const addStep    = () => setDraftSteps(arr => [...arr, ''])
-  const removeStep = i =>
-    setDraftSteps(arr => arr.filter((_, idx) => idx !== i))
-
-  function onFileChange(e) {
-    const file = e.target.files[0]
-    if (file) {
-      setImageFile(file)
-      setDraftImageUrl(URL.createObjectURL(file))
-    }
   }
 
   function parseRecipe(text) {
@@ -240,147 +136,17 @@ export default function Recipes() {
       <div className="cards-grid">
         {filtered.length === 0 && <p>Aucune recette trouvée.</p>}
         {filtered.map(r => (
-          <div key={r.id} className="card" onClick={() => openOverlay(r)}>
-            {r.imageUrl && (
-              <img src={r.imageUrl} alt="" className="card-thumb" />
-            )}
-            <h3 className="card-title">{r.title}</h3>
-          </div>
+          <RecipeCard key={r.id} recipe={r} onSelect={openOverlay} />
         ))}
       </div>
 
       {/* overlay */}
       {selected && (
-        <div
-          className="overlay"
-          onClick={e => {
-            // only close if backdrop clicked
-            if (e.target === e.currentTarget) setSelected(null)
-          }}
-        >
-          <div className="overlay-content" onClick={e => e.stopPropagation()}>
-            <button className="btn-close" onClick={() => setSelected(null)}>✖</button>
-
-            {/* Title */}
-            {editMode
-              ? <input
-                  className="overlay-input-title"
-                  value={draftTitle}
-                  onChange={e => setDraftTitle(e.target.value)}
-                />
-              : <h2 className="overlay-title">{selected.title}</h2>
-            }
-
-            {(editMode ? draftImageUrl : selected.imageUrl) && (
-              <img
-                src={editMode ? draftImageUrl : selected.imageUrl}
-                alt=""
-                className="overlay-image"
-              />
-            )}
-            {editMode && (
-              <div>
-                <input type="file" accept="image/*" onChange={onFileChange} />
-                {draftImageUrl && (
-                  <button className="btn-remove-img" onClick={removeImage}>
-                    Supprimer l'image
-                  </button>
-                )}
-              </div>
-            )}
-
-            {/* Ingredients */}
-            <section className="section">
-              <h3>Ingrédients</h3>
-              {editMode
-                ? <>
-                    {draftIngredients.map((ing,i)=>(
-                      <div key={i} className="list-edit-row">
-                        <input
-                          type="text"
-                          placeholder="Qté…"
-                          value={ing.quantity}
-                          onChange={e=>setIngQty(i, e.target.value)}
-                          style={{ width: '4rem' }}
-                        />
-                        <input
-                          type="text"
-                          placeholder="Nom…"
-                          value={ing.name}
-                          onChange={e=>setIngName(i, e.target.value)}
-                        />
-                        <button onClick={()=>removeIngredient(i)}>🗑️</button>
-                      </div>
-                    ))}
-                    <button className="btn-small" onClick={addIngredient}>
-                      ➕ Ajouter un ingrédient
-                    </button>
-                  </>
-                : selected.ingredients?.length
-                  ? <ul>
-                      {selected.ingredients.map((ing,i)=>(
-                        <li key={i}>{ing.quantity} {ing.name}</li>
-                      ))}
-                    </ul>
-                  : <p>— aucun ingrédient —</p>
-              }
-            </section>
-
-            {/* Steps */}
-            <section className="section">
-              <h3>Préparation</h3>
-              {editMode
-                ? <>
-                    {draftSteps.map((st,i)=>(
-                      <div key={i} className="list-edit-row">
-                        <input
-                          type="text"
-                          placeholder={`Étape ${i+1}…`}
-                          value={st}
-                          onChange={e=>setStep(i, e.target.value)}
-                        />
-                        <button onClick={()=>removeStep(i)}>🗑️</button>
-                      </div>
-                    ))}
-                    <button className="btn-small" onClick={addStep}>
-                      ➕ Ajouter une étape
-                    </button>
-                  </>
-                : selected.steps?.length
-                  ? <ol>
-                      {selected.steps.map((st,i)=><li key={i}>{st}</li>)}
-                    </ol>
-                  : <p>— pas d'étapes —</p>
-              }
-            </section>
-
-            {/* Notes */}
-            <section className="section">
-              <h3>Notes</h3>
-              {editMode
-                ? <textarea
-                    className="overlay-textarea"
-                    value={draftNotes}
-                    onChange={e=>setDraftNotes(e.target.value)}
-                  />
-                : <p>{selected.notes || '— aucune note —'}</p>
-              }
-            </section>
-
-            {/* Footer actions */}
-            <div className="overlay-footer">
-              {editMode
-                ? <button className="btn-save" onClick={saveAll}>💾 Sauvegarder</button>
-                : <button className="btn-modify" onClick={()=>setEditMode(true)}>✏️ Modifier</button>
-              }
-              {!editMode && (
-                <button className="btn-delete" onClick={deleteCurrent}>
-                  🗑️ Supprimer
-                </button>
-              )}
-            </div>
-          </div>
-        </div>
+        <RecipeOverlay
+          recipe={selected}
+          onClose={() => setSelected(null)}
+          onSave={setSelected}
+        />
       )}
 
       {pasteMode && (
